clean up duplication
